`int taint` -> `float flux` // WIP alembic inefficiency

+++ src/main/java/net/arcanamod/client/ClientAuraHandler.java
@@ -29,7 +29,7 @@ public class ClientAuraHandler{
 	// Keep track of loaded chunks on client using events.
 	// Don't keep track of aura at arbitrary positions - the server will update us on the aura at our position.
 	
-	public static int currentFlux = 0;
+	public static float currentFlux = 0;
 	
 	@SubscribeEvent
 	public static void chunkLoadOnClient(ChunkEvent.Load event){
